requirements.txt
datasets==3.1.0
transformers==4.46.2
torch==2.6.0.dev20241107
python-dotenv==1.0.1
datasets==3.1.0
optuna==4.0.0
matplotlib==3.9.2
